virtual_com/pyproject.toml
[project]
name = "virtual-com"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymavlink>=2.4.49",
    "pyserial>=3.5",
]
